+++ src/backend/joanie/tests/core/test_commands_generate_certificates.py
@@ -193,13 +193,13 @@ class CreateCertificatesTestCase(TestCase):
         self.assertEqual(certificate_qs.count(), 0)
 
         # A certificate should be generated for the 1st product
-        with self.assertNumQueries(19):
+        with self.assertNumQueries(23):
             call_command("generate_certificates", product=product_1.id)
         self.assertEqual(certificate_qs.filter(order=orders[0]).count(), 1)
         self.assertEqual(certificate_qs.filter(order=orders[1]).count(), 0)
 
         # Then a certificate should be generated for the 2nd product
-        with self.assertNumQueries(19):
+        with self.assertNumQueries(23):
             call_command("generate_certificates", product=product_2.id)
         self.assertEqual(certificate_qs.filter(order=orders[1]).count(), 1)
 
@@ -296,12 +296,12 @@ class CreateCertificatesTestCase(TestCase):
         self.assertEqual(certificate_qs.count(), 0)
 
         # A certificate should be generated for the 1st product
-        with self.assertNumQueries(19):
+        with self.assertNumQueries(23):
             call_command("generate_certificates", product=product_1.id)
         self.assertEqual(certificate_qs.filter(order=orders[0]).count(), 1)
         self.assertEqual(certificate_qs.filter(order=orders[1]).count(), 0)
 
         # Then a certificate should be generated for the 2nd product
-        with self.assertNumQueries(19):
+        with self.assertNumQueries(23):
             call_command("generate_certificates", product=product_2.id)
         self.assertEqual(certificate_qs.filter(order=orders[1]).count(), 1)