+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Add order export to CSV in back office
+- Branded Unicamp degrees
+
 ### Fixed
 
 - Fix order filtering by creation date
@@ -38,7 +43,6 @@ and this project adheres to
   to `Product`model
 - Add admin api endpoints to CRUD `Teacher` and `Skill` resources.
 - Add certification section in back office product detail view
-- Add order export to CSV in back office
 
 ### Changed