+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.13.0] - 2025-01-15
+
 ### Added
 
 - Add order export to CSV in back office
@@ -549,7 +551,8 @@ and this project adheres to
 - First working version serving sellable micro-credentials for multiple
   organizations synchronized to a remote catalog
 
-[unreleased]: https://github.com/openfun/joanie/compare/v2.12.0...main
+[unreleased]: https://github.com/openfun/joanie/compare/v2.13.0...main
+[2.13.0]: https://github.com/openfun/joanie/compare/v2.12.0...v2.13.0
 [2.12.0]: https://github.com/openfun/joanie/compare/v2.11.0...v2.12.0
 [2.11.0]: https://github.com/openfun/joanie/compare/v2.10.0...v2.11.0
 [2.10.0]: https://github.com/openfun/joanie/compare/v2.9.2...v2.10.0